pass authenticate method instead of "sourceProviderManager"
