--- open-authoring/source-provider-managers/SourceProviderManager.ts
@@ -1,3 +1,0 @@
-export interface SourceProviderManager {
-  authenticate(): Promise<void>
-}
